데이터파일 삭제
데이터파일을 삭제기능 추가
게임 초기화 가능

+++ Assets/Script/Controller/DataController.cs
@@ -1,15 +1,18 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEditor;
+using UnityEngine.SceneManagement;
 using UnityEngine;
 using System.IO;
 
 public class DataController : MonoBehaviour
 {
-    private string DataFile = "Data/database.json";
+    private string DataFile = "database.json";
     private string FilePath;
 
     private Data GameData;
+
+    private bool IsEnd = false;
 
     private static DataController Instance = null;
 
@@ -18,15 +21,9 @@ public class DataController : MonoBehaviour
     // 게임데이터 삭제
     public void DeleteGameData()
     {
-        FileInfo file_info = new FileInfo(FilePath);
+        if (File.Exists(FilePath)) File.Delete(FilePath);
+        IsEnd = !IsEnd;
 
-        if(File.Exists(FilePath))
-        {
-            Debug.Log(FilePath + " /// " + DataFile + " /// " + GameData);
-            //FileUtil.DeleteFileOrDirectory(FilePath);
-            //file_info.Delete();
-            File.Delete(FilePath);
-        }
         GameObject.Find("GameController").GetComponent<GameController>().ExitGame();
     }
 
@@ -35,14 +32,11 @@ public class DataController : MonoBehaviour
     {
         string jsonData = JsonUtility.ToJson(GameData);
         File.WriteAllText(FilePath, jsonData);
-        Debug.Log(jsonData);
-
     }
 
     // 새로운 게임데이터 생성
     public void NewGameData()
     {
-        Debug.Log("새게임 생성");
         GameData = new Data();
     }
 
@@ -55,28 +49,29 @@ public class DataController : MonoBehaviour
             GameData = JsonUtility.FromJson<Data>(jsonData);
         }
         else NewGameData();
-
     }
 
     private void OnApplicationQuit()
     {
-        SaveGameData();
+        if (!IsEnd) SaveGameData();
     }
 
     private void Awake()
     {
         // 안드로이드용 패스
-        //FilePath = Application.persistentDataPath + DataFile;
+        FilePath = Application.persistentDataPath + DataFile;
         // PC용 패스 ( 테스트용 )
-        FilePath = Path.Combine(Application.dataPath, DataFile);
+        //FilePath = Path.Combine(Application.dataPath, DataFile);
 
+        
         if (Instance)
         {
             DestroyImmediate(gameObject);
             return;
         }
         Instance = this;
-        DontDestroyOnLoad(gameObject);
+        DontDestroyOnLoad(gameObject);        
+        
     }
 
     void Start()
@@ -86,6 +81,6 @@ public class DataController : MonoBehaviour
 
     void Update()
     {
-        if (GameData == null) LoadGameData();
+        if (GameData == null && !IsEnd) LoadGameData();
     }
 }
